reworking gallery fragment, with view record fragment
