+++ app/build.gradle
@@ -65,9 +65,11 @@ dependencies {
     // Navigation Component
     implementation "androidx.navigation:navigation-fragment-ktx:2.4.1"
     implementation "androidx.navigation:navigation-ui-ktx:2.4.1"
+    implementation 'com.squareup.picasso:picasso:2.71828'
 
     implementation 'androidx.cardview:cardview:1.0.0'
-
+    implementation 'com.google.firebase:firebase-storage-ktx:20.1.0'
+    implementation 'com.google.firebase:firebase-firestore-ktx:24.5.0'
     implementation platform('com.google.firebase:firebase-bom:31.4.0')
     implementation 'com.google.firebase:firebase-firestore-ktx'
 }
