--- app/src/main/java/com/example/fishbook/Species.kt
@@ -1,15 +1,0 @@
-package com.example.fishbook
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import java.util.UUID
-
-
-@Entity(tableName = "species_table")
-data class Species(
-    @PrimaryKey val id: UUID = UUID.randomUUID(),
-    val caught_flag: Boolean,
-    val species_name: String,
-    val fish_family: String,
-    val image: Int = 0
-)
-
