--- app/src/main/java/com/example/fishbook/SpeciesApplication.kt
@@ -1,10 +1,0 @@
-package com.example.fishbook
-
-import android.app.Application
-
-class SpeciesApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        SpeciesRepository.initialize(this)
-    }
-}
