+++ app/src/main/java/com/example/fishbook/fishdex/SpeciesApplication.kt
@@ -1,0 +1,10 @@
+package com.example.fishbook.fishdex
+
+import android.app.Application
+
+class SpeciesApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        SpeciesRepository.initialize(this)
+    }
+}
