+++ app/src/main/java/com/example/fishbook/gallery/Gallery.kt
@@ -1,0 +1,65 @@
+package com.example.fishbook.gallery
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.viewModels
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
+import androidx.navigation.fragment.findNavController
+
+import com.example.fishbook.databinding.FragmentGalleryBinding
+import kotlinx.coroutines.launch
+
+
+/**
+ * A simple [Fragment] subclass.
+ * Use the [Gallery.newInstance] factory method to
+ * create an instance of this fragment.
+ */
+
+class Gallery : Fragment() {
+    private var _binding: FragmentGalleryBinding? = null
+    private val binding get() = _binding!!
+
+    private val galleryViewModel: GalleryViewModel by viewModels()
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentGalleryBinding.inflate(inflater, container, false)
+
+        /*binding.uploadbutton.setOnClickListener {
+            val intent = Intent(activity, storageActivity::class.java)
+            startActivity(intent)
+        }*/
+        return binding.root
+
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        viewLifecycleOwner.lifecycleScope.launch {
+            viewLifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                galleryViewModel.GridImages.collect { gridImages ->
+                    binding.gridView.adapter =
+                        GalleryAdapter(requireContext(), gridImages) { gridImage ->
+                            findNavController().navigate(
+                                GalleryDirections.showRecord(gridImage)
+                            )
+                        }
+                }
+            }
+        }
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
