+++ app/src/main/java/com/example/fishbook/species_database/SpeciesDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.example.fishbook.Species
+import com.example.fishbook.fishdex.Species
 import kotlinx.coroutines.flow.Flow
 import java.util.*
