+++ app/src/main/java/com/example/fishbook/species_database/SpeciesDatabase.kt
@@ -3,7 +3,7 @@ package com.example.fishbook.species_database
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.example.fishbook.Species
+import com.example.fishbook.fishdex.Species
 
 @Database(entities = [Species::class], version = 1, exportSchema = false)
 @TypeConverters(SpeciesTypeConverters::class)
