+++ app/src/main/res/layout/fragment_fish_dex.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FishDexFragment">
+    tools:context=".fishdex.FishDexFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/fishRecyclerView"
