+++ app/src/main/res/layout/fragment_gallery.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Gallery">
+    tools:context=".gallery.Gallery">
 
     <!-- TODO: Update blank fragment layout -->
     <LinearLayout
